# openapi.yaml
openapi: 3.1.0
info:
  title: Smart Home Symcon Service
  description: API сервис управления устройствами умного дома.
  contact:
    name: Alexey Kuznetsov
    url: https://github.com/alexeycoder
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://localhost:8080/v1
    description: Test server
  - url: https://symcon.ru/v1
    description: Production server
...
paths:
...
  /devices:
    post:
      tags:
        - devices
      summary: Добавить устройство
      description: Добавить устройство
      operationId: addDevice
      parameters:
        - name: groupId
          in: header
          description: Идентификатор группы
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Создание нового устройства в базе
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
        required: true
      responses:
        '201':
          description: Created successfully
        '400':
          description: Invalid device data supplied
        '401':
          description: Unauthorized
  /devices/{deviceId}:
    put:
      tags:
        - devices
      summary: Обновить робота
      description: Обновить робота
      operationId: updateRobot
      parameters:
        - name: groupId
          in: header
          description: Идентификатор группы
          required: true
          schema:
            type: integer
            format: int64
        - name: deviceId
          in: path
          description: Идентификатор устройства
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Обновить данные устройства в базе
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Device not found

    delete:
      tags:
        - devices
      summary: Удаляет робота
      description: Удаляет робота
      operationId: deleteRobot
      parameters:
        - name: groupId
          in: header
          description: Идентификатор группы
          required: true
          schema:
            type: integer
            format: int64
        - name: deviceId
          in: path
          description: Идентификатор устройства
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
  /devices/{deviceId}/schedules:
    get:
      tags:
        - schedules
      summary: Добавить график работ
      description: Добавить график работ
      operationId: findAllSchedules
      parameters:
        - name: deviceId
          in: path
          description: Идентификатор устройства
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized

    post:
      tags:
        - schedules
      summary: Добавить график работ
      description: Добавить график работ
      operationId: addSchedule
      parameters:
        - name: deviceId
          in: path
          description: Идентификатор устройства
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Создание нового устройства в базе
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
        required: true
      responses:
        '201':
          description: Created successfully
        '400':
          description: Invalid ID or data supplied
        '401':
          description: Unauthorized
...
components:
  schemas:
...
    Device:
      required:
        - typeId
        - name
        - modelId
        - serialNum
        - firmwareVersion
        - connectionId
      type: object
      properties:
        deviceId:
          type: integer
          format: int64
          examples:
            - 123
        typeId:
          type: integer
          format: int32
          examples:
            - 1
            - 2
          description: Код производителя типа устройства, например, 1 - робот-пылесос
        name:
          type: string
          examples:
            - Робот Кузя
        modelId:
          type: integer
          format: int32
        serialNum:
          type: string
          examples:
            - "6a0fa287-6969-4839"
        firmwareVersion:
          type: integer
          format: int32
          examples:
            - 2
        connectionId:
          type: integer
          format: int64
          examples:
            - 123
...
    Schedule:
      required:
        - deviceId
        - dayOfWeek
        - timeOfDay
        - serviceTypeId
        - servicePropertiesCode
      type: object
      properties:
        scheduleId:
          type: integer
          format: int64
          examples:
            - 123
        dayOfWeek:
          type: integer
          format: int32
        timeOfDay:
          type: string
          format: time
          examples:
            - "10:35"
        serviceTypeId:
          type: integer
          format: int32
        servicePropertiesCode:
          type: integer
          format: int64
          examples:
            - 0xFFFFFFFF