# spring-server/src/main/java/io/swagger/api/DevicesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.62).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Device;
import io.swagger.model.Schedule;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2024-09-27T16:36:05.325149630Z[GMT]")
@Validated
public interface DevicesApi {

    @Operation(summary = "Добавить устройство", description = "Добавить устройство", tags={ "devices" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "201", description = "Created successfully"),
        
        @ApiResponse(responseCode = "400", description = "Invalid device data supplied"),
        
        @ApiResponse(responseCode = "401", description = "Unauthorized") })
    @RequestMapping(value = "/devices",
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> addDevice(@Parameter(in = ParameterIn.HEADER, description = "Идентификатор группы" ,required=true,schema=@Schema()) @RequestHeader(value="groupId", required=true) Object groupId
, @Parameter(in = ParameterIn.DEFAULT, description = "Создание нового устройства в базе", required=true, schema=@Schema()) @Valid @RequestBody Device body
);


    @Operation(summary = "Добавить график работ", description = "Добавить график работ", tags={ "schedules" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "201", description = "Created successfully"),
        
        @ApiResponse(responseCode = "400", description = "Invalid ID or data supplied"),
        
        @ApiResponse(responseCode = "401", description = "Unauthorized") })
    @RequestMapping(value = "/devices/{deviceId}/schedules",
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> addSchedule(@Parameter(in = ParameterIn.PATH, description = "Идентификатор устройства", required=true, schema=@Schema()) @PathVariable("deviceId") Object deviceId
, @Parameter(in = ParameterIn.DEFAULT, description = "Создание нового устройства в базе", required=true, schema=@Schema()) @Valid @RequestBody Schedule body
);


    @Operation(summary = "Удаляет робота", description = "Удаляет робота", tags={ "devices" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "204", description = "Deleted successfully"),
        
        @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
        
        @ApiResponse(responseCode = "401", description = "Unauthorized") })
    @RequestMapping(value = "/devices/{deviceId}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteRobot(@Parameter(in = ParameterIn.HEADER, description = "Идентификатор группы" ,required=true,schema=@Schema()) @RequestHeader(value="groupId", required=true) Object groupId
, @Parameter(in = ParameterIn.PATH, description = "Идентификатор устройства", required=true, schema=@Schema()) @PathVariable("deviceId") Object deviceId
);


    @Operation(summary = "Добавить график работ", description = "Добавить график работ", tags={ "schedules" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))),
        
        @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
        
        @ApiResponse(responseCode = "401", description = "Unauthorized") })
    @RequestMapping(value = "/devices/{deviceId}/schedules",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Object> findAllSchedules(@Parameter(in = ParameterIn.PATH, description = "Идентификатор устройства", required=true, schema=@Schema()) @PathVariable("deviceId") Object deviceId
);


    @Operation(summary = "Обновить робота", description = "Обновить робота", tags={ "devices" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK"),
        
        @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
        
        @ApiResponse(responseCode = "401", description = "Unauthorized"),
        
        @ApiResponse(responseCode = "404", description = "Device not found") })
    @RequestMapping(value = "/devices/{deviceId}",
        consumes = { "application/json" }, 
        method = RequestMethod.PUT)
    ResponseEntity<Void> updateRobot(@Parameter(in = ParameterIn.HEADER, description = "Идентификатор группы" ,required=true,schema=@Schema()) @RequestHeader(value="groupId", required=true) Object groupId
, @Parameter(in = ParameterIn.PATH, description = "Идентификатор устройства", required=true, schema=@Schema()) @PathVariable("deviceId") Object deviceId
, @Parameter(in = ParameterIn.DEFAULT, description = "Обновить данные устройства в базе", required=true, schema=@Schema()) @Valid @RequestBody Device body
);

}